requirements.txt
web3==6.15.1
python-dotenv==1.0.1
eth-account==0.11.0
